Added checks for if a user has a balance as well as changes to calculations
user now cannot have a balance when they hit checkout... they have to resolve all of the overdue rentals associated with their account before they can checkout with the items in their cart.

Moved the Rental.IsLate Calculations into the Login function, so when a user logs in, all due dates and prices are calculated and set.

also changes the check balance page to include the list of rented movies and a redirect to the return rental page